NUM-167 Move AmqpConsumerRequest and AmqpConsumerResponse to api.consumer
