+++ amqp-driver/src/main/scala/com/nummulus/amqp/driver/api/consumer/AmqpConsumerResponse.scala
@@ -1,3 +1,3 @@
-package com.nummulus.amqp.driver.akka
+package com.nummulus.amqp.driver.api.consumer
 
 case class AmqpConsumerResponse(body: String)
